# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AnubisWatch API
  description: |
    AnubisWatch REST API - The Judgment Never Sleeps

    This API provides full access to AnubisWatch monitoring capabilities including:
    - Soul (monitor) management
    - Judgment (check) history
    - Alert channels and rules
    - Status pages
    - Cluster management

    ## Authentication

    All API endpoints except `/health`, `/ready`, and `/api/v1/auth/login` require Bearer token authentication.

    ### Getting a Token

    ```bash
    curl -X POST http://localhost:8443/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email":"admin@example.com","password":"secret"}'
    ```

    ### Using the Token

    ```bash
    curl http://localhost:8443/api/v1/souls \
      -H "Authorization: Bearer YOUR_TOKEN"
    ```

    ## Base URL

    - Default: `http://localhost:8443`
    - Production: Configure via `ANUBIS_HOST` and `ANUBIS_PORT` environment variables
  version: 1.0.0
  contact:
    name: AnubisWatch
    url: https://github.com/AnubisWatch/anubiswatch
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8443
    description: Local development server
  - url: https://localhost:8443
    description: Local development server (HTTPS)
  - url: https://anubis.example.com
    description: Production server
...
paths:
...
  /api/v1/souls/{id}/check:
    post:
      tags:
        - Souls
      summary: Force check
      description: Trigger an immediate check for a soul
      operationId: forceCheck
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Judgment'
...
  /api/v1/incidents/{id}/acknowledge:
    post:
      tags:
        - Incidents
      summary: Acknowledge incident
      description: Acknowledge an incident
      operationId: acknowledgeIncident
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Incident acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: acknowledged
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token received from /api/v1/auth/login
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
...
  schemas:
...
    Judgment:
      type: object
      properties:
        id:
          type: string
        soul_id:
          type: string
        status:
          type: string
          enum: [alive, degraded, dead, unknown]
        latency_ms:
          type: number
        error:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          $ref: '#/components/schemas/JudgmentDetails'
    JudgmentDetails:
      type: object
      properties:
        status_code:
          type: integer
        tls_info:
          $ref: '#/components/schemas/TLSInfo'
        assertions:
          type: array
          items:
            $ref: '#/components/schemas/AssertionResult'
    TLSInfo:
      type: object
      properties:
        version:
          type: string
        cipher_suite:
          type: string
        expiry:
          type: string
          format: date-time
        issuer:
          type: string
    AssertionResult:
      type: object
      properties:
        name:
          type: string
        passed:
          type: boolean
        message:
          type: string
...